Convert channel inputs from strings to explicit type.
The documentation and examples are scarce on which strings need to be provided
as channel identifiers. To allow better explorability of respective channels
and to assure that the strings that are transfered to the TCP server are
correct, an Enumeration type is introduced. Its values can be used to provide
channel identifiers to the enable_channel method of the Arc class.

See the adapted examples on the effect of this change.