Refactor enable/disable operations on channels to toggling API.
Previously, the API contained surprises to users, such as, calls like
`enable_main(False)`, which would disable the main power of a device.
Additionaly, in the earlier version the API was inconsistent. Some things were
enabled (via methods like `enable_*(enable)`, where `enable` was a boolean
value) or they were set (via methods like `set_*(enable)`, where `enable` was
a boolean a value).

This commit introduces a new enumeration type `Toggle` in toggle.py and
refactors the API in the `Arc` class and respective examples accordingly. The
refactoring harmonizes the API. Each channel or functionality that can be
toggle has now a corresponding `toggle_*` method in the `Arc` class. Next to it
there are two corresponding methods that `enable_*` or `disable_*`
the respective channel.

That is, this commit make the APIU more explicit and harmonizes it.